Add the setattr_x callback to list of supported functions
This one is specific to OSX (and OSX FUSE supports it).
It gives the caller the ability to do multiple changes at once.

Without having implemented this callback, apps like TextEdit or Preview
won't save changes to the file. The problem is probably triggered by the
use of the NSDocument
class (https://developer.apple.com/documentation/appkit/nsdocument)
which provides the feature to safely write changes to the underlying file
(it copies the file, saves changes and then does a rename operation).

This callback is not present in the windows version of the API (the
setattr_x struct from fuse_common.h is not even defined).